refactor: align debug symbols with command families
