+++ src/__tests__/debug-symbols.test.ts
@@ -3,7 +3,7 @@ import fs from 'node:fs/promises';
 import os from 'node:os';
 import path from 'node:path';
 import { test } from 'vitest';
-import { symbolicateCrashArtifact } from '../debug-symbols.ts';
+import { symbolicateCrashArtifact } from '../commands/debugging/runtime/debug-symbols.ts';
 import { AppError } from '../utils/errors.ts';
 import { withCommandExecutorOverride } from '../utils/exec.ts';
