+++ src/commands/client-command-facets.ts
@@ -1,3 +1,4 @@
+import { debuggingCommandDefinitions, debuggingCommandMetadata } from './debugging/index.ts';
 import { captureCommandDefinitions, captureCommandMetadata } from './capture/index.ts';
 import { managementCommandDefinitions, managementCommandMetadata } from './management/index.ts';
 import { metroCommandDefinition, metroCommandMetadata } from './metro/index.ts';
@@ -5,6 +6,7 @@ import {
   observabilityCommandDefinitions,
   observabilityCommandMetadata,
 } from './observability/index.ts';
+import { perfCommandDefinitions, perfCommandMetadataList } from './perf/index.ts';
 import { reactNativeCommandDefinition, reactNativeCommandMetadata } from './react-native/index.ts';
 import { recordingCommandDefinitions, recordingCommandMetadata } from './recording/index.ts';
 import { replayCommandDefinitions, replayCommandMetadataList } from './replay/index.ts';
@@ -34,6 +36,14 @@ const clientCommandFamilyFacets = [
   {
     metadata: observabilityCommandMetadata,
     definitions: observabilityCommandDefinitions,
+  },
+  {
+    metadata: perfCommandMetadataList,
+    definitions: perfCommandDefinitions,
+  },
+  {
+    metadata: debuggingCommandMetadata,
+    definitions: debuggingCommandDefinitions,
   },
   {
     metadata: recordingCommandMetadata,
