+++ src/commands/perf/perf-command-contract.ts
